Folder reorganization - UI component moved to /shared
