+++ quick-start/src/main/ui/app/jobs/index.ts
@@ -1,2 +1,5 @@
+export * from './jobs.component';
+export * from './job-output.component';
+export * from './job-export.component';
 export * from './jobs.service';
 export * from './job-listener.service';
