+++ quick-start/src/main/ui/app/jobs/job-export.component.ts
@@ -1,6 +1,6 @@
 import {Component, HostListener, Inject, Input} from '@angular/core';
 import {MdlDialogReference, MdlDialogService} from '@angular-mdl/core';
-import {JobService} from "../../../jobs";
+import {JobService} from "./jobs.service";
 
 @Component({
   selector: 'job-export-dialog',
