+++ quick-start/src/main/ui/app/jobs/job-output.component.html
