+++ quick-start/src/main/ui/app/jobs/job.model.ts
