+++ quick-start/src/main/ui/app/jobs/jobs.component.README.md
