+++ quick-start/src/main/ui/app/jobs/jobs.component.scss
@@ -1,4 +1,4 @@
-@import '../theme/styles/variables';
+@import '../shared/components/theme/styles/variables';
 
 $bg-color: unquote("rgb(#{$palette-datahub-500})");
