+++ quick-start/src/main/ui/app/shared/components/jobs/index.ts
@@ -1,4 +1,1 @@
-export * from './jobs.component';
 export * from './jobs-ui.component';
-export * from './job-output.component';
-export * from './job-export.component';
